empty state overlay array buffer tests
